==> fra/12_fra_explore.Rmd <==
---
title: "12_fra_explore"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)

library(tidyverse)
library(scales)
library(ggunchained)
library(lmerTest)
```

# CHECK SHUFFLED DATA
Load in, plot, save
```{r}
ntr_paths = list.files("fra/data/ntr_out/", pattern = "*.csv", full.names = T)
shuff_ntr_paths = list.files("fra/data/ntr_shuffled/", pattern = "*.csv", full.names = T)


export_plots <- function(path) {
  
  # load in
  ntr_w = read_csv(path)
  
  # extract timeframe
  w = str_extract(path, "(?<=W)\\d+")
  
  
  # RESONANCE ~ NOVELTY
  ggplot(ntr_w, aes(z_novelty, z_resonance)) +
    geom_bin2d(bins = 100) +
    scale_fill_viridis_c(trans = "log", breaks = c(1,10,100,1000),
                         labels = trans_format("log10", 
                                               math_format(expr = 10^.x, format = force))) +
    geom_smooth(method = "lm", colour = "black", alpha = 0, size = 0.5) +
    labs(x = "Novelty (z-scaled)\n", y = "Resonance (z-scaled)", 
         title = paste0("Shuffled Resonance vs. Novelty, w = ", w),
         subtitle = "with a regression line",
         caption = "Data source: French National Assembly debates (1998-2019)",
         fill = "speech\ncount") +
    scale_x_continuous(breaks = seq(-4, 8, 2)) +
    scale_y_continuous(breaks = seq(-6, 6, 2)) +
    theme_janco_point() +
    theme(legend.direction = "vertical", legend.position = "right",
          plot.caption = element_text(hjust = 0.5)
    )
  
  ggsave(filename = paste0("shuff_rn_", w, ".png"), path = "fra/data/plots/shuffled/")
  
  
  
  # NOVELTY VS TRANSIENCE
  ggplot(ntr_w, aes(z_novelty, z_transience)) +
    geom_bin2d(bins = 100) +
    scale_fill_viridis_c(trans = "log", breaks = c(1,10,100,1000,10000),
                         labels = trans_format("log10", math_format(expr = 10^.x, format = force))) +
    geom_abline(colour = "black", linetype = "dashed") +
    labs(x = "Novelty (z-scaled)\n", y = "Transience (z-scaled)", 
         title = paste0("Shuffled Novelty vs. Transience, w =", w),
         subtitle = "with an identity line (x = y)",
         caption = "Data source: French National Assembly debates (1998-2019)",
         fill = "speech\ncount") +
    scale_x_continuous(breaks = seq(-4, 8, 2)) +
    scale_y_continuous(breaks = seq(-6, 8, 2)) +
    theme_janco_point() +
    theme(legend.direction = "vertical", legend.position = "right",
          plot.caption = element_text(hjust = 0.5)
    )
  
  ggsave(filename = paste0("shuff_nt_", w, ".png"), path = "fra/data/plots/shuffled/")
  
  
}

lapply(shuff_ntr_paths, export_plots)
```

See the damn shuffled dataset
```{r}
shuff_41 = read_csv(shuff_ntr_paths[3]) %>%
  rename(rowname = doc_id)

ref = read_csv("fra/data/03_fra_ref.csv")

shuff_41 = left_join(shuff_41, ref, by = "rowname")
```

Did it shuffle correctly?
YES
```{r}
fra_full = read_csv("fra/data/03_fra_s.csv")

a = shuff_41 %>%
  filter(rowname == 891285)

b = fra_full %>%
  filter(rowname == 891285)
```

Modelling the difference?
```{r}
ntr_41 = read_csv("fra/data/ntr_out/W41__from_matrix.csv") %>%
  rename(rowname = doc_id)

ntr_41 = left_join(ntr_41, ref, by = "rowname")

norm = lm(ntr_41$resonance ~ ntr_41$novelty)
summary(norm)

shuff = lm(shuff_41$resonance ~ shuff_41$novelty)
summary(shuff)
```


Var distribution
```{r}
path_norm = "fra/data/plots/dist/norm_"

rethinking::dens(ntr_41$novelty) 
rethinking::dens(ntr_41$transience)
rethinking::dens(ntr_41$resonance)

rethinking::dens(shuff_41$novelty)
rethinking::dens(shuff_41$transience)
rethinking::dens(shuff_41$resonance)
```


RE
```{r}
shuff_41_same = shuff_41 %>%
  slice(1:nrow(ntr_41))

norm_r1 = lmer(z_transience ~ z_novelty + (1 | name), data = ntr_41, REML = FALSE)
shuff_r1 = lmer(z_transience ~ z_novelty + (1 | name), data = shuff_41_same, REML = FALSE)

summary(norm_r1)
summary(shuff_r1)
```

# new
```{r}
norm41 = read_csv(ntr_paths[str_detect(ntr_paths, "41")])
shuff41 = read_csv(shuff_ntr_paths[str_detect(shuff_ntr_paths, "41")])
```

```{r}
shuff41 %>%
ggplot(aes(z_transience, z_novelty)) +
    geom_bin2d(bins = 100) +
    scale_fill_viridis_c(trans = "log", breaks = c(1,10,100,1000,10000),
                         labels = trans_format("log10", math_format(expr = 10^.x, format = force))) +
    scale_x_continuous(breaks = seq(-4, 8, 2)) +
    scale_y_continuous(breaks = seq(-6, 8, 2)) +
    theme_janco_point() +
    theme(legend.direction = "vertical", legend.position = "right",
          plot.caption = element_text(hjust = 0.5)
    )

ggsave("fra/data/plots/minimal/tn.png")


shuff41 %>%
ggplot(aes(z_novelty, z_transience)) +
    geom_bin2d(bins = 100) +
    scale_fill_viridis_c(trans = "log", breaks = c(1,10,100,1000,10000),
                         labels = trans_format("log10", math_format(expr = 10^.x, format = force))) +
    scale_x_continuous(breaks = seq(-4, 8, 2)) +
    scale_y_continuous(breaks = seq(-6, 8, 2)) +
    theme_janco_point() +
    theme(legend.direction = "vertical", legend.position = "right",
          plot.caption = element_text(hjust = 0.5)
    )
ggsave("fra/data/plots/minimal/nt.png")


shuff41 %>%
  ggplot(aes(z_transience, z_resonance)) +
    geom_bin2d(bins = 100) +
    scale_fill_viridis_c(trans = "log", breaks = c(1,10,100,1000,10000),
                         labels = trans_format("log10", math_format(expr = 10^.x, format = force))) +
    scale_x_continuous(breaks = seq(-4, 8, 2)) +
    scale_y_continuous(breaks = seq(-6, 8, 2)) +
    theme_janco_point() +
    theme(legend.direction = "vertical", legend.position = "right",
          plot.caption = element_text(hjust = 0.5)
    )
ggsave("fra/data/plots/minimal/tr.png")


shuff41 %>%
  ggplot(aes(z_novelty, z_resonance)) +
    geom_bin2d(bins = 100) +
    scale_fill_viridis_c(trans = "log", breaks = c(1,10,100,1000,10000),
                         labels = trans_format("log10", math_format(expr = 10^.x, format = force))) +
    scale_x_continuous(breaks = seq(-4, 8, 2)) +
    scale_y_continuous(breaks = seq(-6, 8, 2)) +
    theme_janco_point() +
    theme(legend.direction = "vertical", legend.position = "right",
          plot.caption = element_text(hjust = 0.5)
    )
ggsave("fra/data/plots/minimal/nr.png")
```
